Implement parallel raytracing via Rayon
This change speeds up the final image of book 1 from ~7000 seconds to
~255-270 seconds on my development machine. Instead of calculating each
scanline in parallel, each pixel's color is calculated in parallel, and
each sample for each pixel is also parallelized.

The info logging to track progress was updated to run on a separate
thread. This would be a good option for the camera to disable in the
future to increase speed further.